video_stm_activity/requirements.txt
torch
torchvision
transformers>=4.41.0
accelerate
bitsandbytes
opencv-python
pandas
pyyaml
tqdm
scikit-learn
Pillow
matplotlib
seaborn
pandas
